+++ chemcloud/exceptions.py
@@ -4,3 +4,11 @@ class BaseError(Exception):
 
 class TimeoutError(BaseError):
     """A timeout parameter was exceeded"""
+
+
+class UnsupportedProgramError(BaseError):
+    """A program is not supported by ChemCloud"""
+
+
+class AuthenticationError(BaseError):
+    """An error occurred during authentication."""